Revert "Remove need for install script"
