+++ lib/commands.js
@@ -19,10 +19,9 @@ var fs = require("fs"),
     deprecator = utils.deprecator,
     asserters = require("./asserters"),
     Asserter = asserters.Asserter,
-    browserScripts = require('./browser-scripts'),
-    safeExecuteJsScript = browserScripts.safeExecuteJsScript,
-    safeExecuteAsyncJsScript = browserScripts.safeExecuteAsyncJsScript,
-    _waitForConditionInBrowserJsScript = browserScripts.waitForConditionInBrowserJsScript;
+    safeExecuteJsScript = require('../build/safe-execute'),
+    safeExecuteAsyncJsScript = require('../build/safe-execute-async'),
+    _waitForConditionInBrowserJsScript = require('../build/wait-for-cond-in-browser');
 
 var commands = {};
