+++ package.json
@@ -80,6 +80,7 @@
   },
   "scripts": {
     "test": "gulp test-unit",
-    "all_tests": "gulp test"
+    "all_tests": "gulp test",
+    "install": "node scripts/build-browser-scripts"
   }
 }
